Further implementation of hotness rating to posts.

+++ db/migrate/20180917035606_add_hotness_to_posts.rb
@@ -1,5 +1,5 @@
 class AddHotnessToPosts < ActiveRecord::Migration[5.2]
   def change
-    add_column :posts, :hotness, :integer
+    add_column :posts, :hotness, :float
   end
 end
